Fix BaseCache.contains() for multipart POST requests
